refactor: integrate wagmi configuration into Nuxt plugin and remove redundant config file
